// src/app/shared/store/types/manufacturing.type.ts
export enum ManufacturingActionTypes {
  INITIALIZE_MANUFACTURING_OPTIONS = 'INITIALIZE_MANUFACTURING_OPTIONS',

  GET_MANUFACTURING_TYPES = 'GET_MANUFACTURING_TYPES',
  GET_MANUFACTURING_TYPES_SUCCESS = 'GET_MANUFACTURING_TYPES_SUCCESS',
  GET_MANUFACTURING_TYPES_FAILURE = 'GET_MANUFACTURING_TYPES_FAILURE',

  GET_STORAGE_TEMPERATURES = 'GET_STORAGE_TEMPERATURES',
  GET_STORAGE_TEMPERATURES_SUCCESS = 'GET_STORAGE_TEMPERATURES_SUCCESS',
  GET_STORAGE_TEMPERATURES_FAILURE = 'GET_STORAGE_TEMPERATURES_FAILURE',

  GET_SHIPPING_CONDITIONS = 'GET_SHIPPING_CONDITIONS',
  GET_SHIPPING_CONDITIONS_SUCCESS = 'GET_SHIPPING_CONDITIONS_SUCCESS',
  GET_SHIPPING_CONDITIONS_FAILURE = 'GET_SHIPPING_CONDITIONS_FAILURE',

  GET_MANUFACTURING_TYPE_INFO = 'GET_MANUFACTURING_TYPE_INFO',
  GET_MANUFACTURING_TYPE_INFO_SUCCESS = 'GET_MANUFACTURING_TYPE_INFO_SUCCESS',
  GET_MANUFACTURING_TYPE_INFO_FAILURE = 'GET_MANUFACTURING_TYPE_INFO_FAILURE',

  UPDATE_MANUFACTURING_MODEL = 'UPDATE_MANUFACTURING_MODEL',

  SET_PRODUCT_TYPE = 'SET_PRODUCT_TYPE',

  RESET_MANUFACRURING_STORE = 'RESET_MANUFACRURING_STORE',

  RESET_MANUFACRURING_CHILDREN = 'RESET_MANUFACRURING_CHILDREN',
  GET_DROPDOWN_LIST_SUCCESS_MANUFACTURING = 'GET_DROPDOWN_LIST_SUCCESS_MANUFACTURING',
  GET_DROPDOWN_LIST_FAILURE_MANUFACTURING = 'GET_DROPDOWN_LIST_FAILURE_MANUFACTURING',

  CLEAR_STATE = 'CLEAR_STATE'
}